# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for ordering food.",
        "title": "Food Ordering API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/promotions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลรายละเอียดของโปรโมชั่นทั้งหมด",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotions"
                ],
                "summary": "ดึงข้อมูลโปรโมชั่นAll",
                "responses": {
                    "200": {
                        "description": "รายละเอียดของโปรโมชั่น",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.PromotionResponse"
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "ไม่พบโปรโมชั่นที่ระบุ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "สร้างโปรโมชั่นใหม่พร้อมรายการสินค้าที่ร่วมรายการ สามารถเป็นได้ทั้งแบบส่วนลดและแบบบันเดิล",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotions"
                ],
                "summary": "สร้างโปรโมชั่นใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลโปรโมชั่น",
                        "name": "promotion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_handlers.createPromo_req"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "รายละเอียดของโปรโมชั่นที่สร้างสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.PromotionResponse"
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "ไม่มีสิทธิ์เข้าถึง",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/promotions/Active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงรายการโปรโมชั่นทั้งหมดที่กำลังใช้งานอยู่ (IsActive=true) และอยู่ในช่วงวันที่ที่กำหนด",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotions"
                ],
                "summary": "ดึงรายการโปรโมชั่นที่กำลังใช้งาน",
                "responses": {
                    "200": {
                        "description": "รายการโปรโมชั่นที่กำลังใช้งาน",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_handlers.PromotionResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "ไม่มีสิทธิ์เข้าถึง",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/promotions/status/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "เปิดหรือปิดการใช้งานโปรโมชั่น โดยการอัพเดทค่า IsActive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotions"
                ],
                "summary": "อัพเดทสถานะการใช้งานโปรโมชั่น",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ของโปรโมชั่น",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "สถานะใหม่ของโปรโมชั่น",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_handlers.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "อัพเดทสถานะสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "ไม่มีสิทธิ์เข้าถึง",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "ไม่พบโปรโมชั่นที่ระบุ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/promotions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงข้อมูลรายละเอียดของโปรโมชั่นตาม ID ที่ระบุ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotions"
                ],
                "summary": "ดึงข้อมูลโปรโมชั่นตาม ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ของโปรโมชั่น",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "รายละเอียดของโปรโมชั่น",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.PromotionResponse"
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "ไม่พบโปรโมชั่นที่ระบุ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "อัพเดตข้อมูลโปรโมชั่นและรายการสินค้าที่ร่วมรายการ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotions"
                ],
                "summary": "อัพเดตข้อมูลโปรโมชั่น",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ของโปรโมชั่น",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ข้อมูลโปรโมชั่นที่ต้องการอัพเดต",
                        "name": "promotion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_handlers.updatePromo_req"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "รายละเอียดของโปรโมชั่นที่อัพเดตสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.PromotionResponse"
                        }
                    },
                    "400": {
                        "description": "ข้อมูลไม่ถูกต้อง",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "ไม่พบโปรโมชั่นที่ระบุ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ลบโปรโมชั่นแบบ Soft Delete โดยการตั้งค่า DeletedAt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotions"
                ],
                "summary": "ลบโปรโมชั่น (Soft Delete)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ของโปรโมชั่น",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ลบโปรโมชั่นสำเร็จ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "ไม่ได้รับอนุญาต",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "ไม่พบโปรโมชั่นที่ระบุ",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "เกิดข้อผิดพลาดภายในเซิร์ฟเวอร์",
                        "schema": {
                            "$ref": "#/definitions/api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input format"
                }
            }
        },
...
        "api_handlers.PromotionItemResponse": {
            "type": "object",
            "properties": {
                "bundle_price": {
                    "type": "number"
                },
                "discount_amount": {
                    "type": "number"
                },
                "discount_percent": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "menu_item": {
                    "$ref": "#/definitions/api_handlers.MenuItemBasic"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "api_handlers.PromotionResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api_handlers.PromotionItemResponse"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "ลด 20% ทุกเมนูข้าว"
                },
                "nameCh": {
                    "type": "string"
                },
                "nameEn": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "example": "single_discount"
                }
            }
        },
...
        "api_handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Status updated successfully"
                }
            }
        },
...
        "api_handlers.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "api_handlers.createPromo_req": {
            "type": "object",
            "required": [
                "items",
                "name",
                "price"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "menu_item_id",
                            "quantity"
                        ],
                        "properties": {
                            "menu_item_id": {
                                "type": "integer"
                            },
                            "quantity": {
                                "type": "integer",
                                "minimum": 1
                            }
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "nameCh": {
                    "type": "string"
                },
                "nameEn": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "api_handlers.updatePromo_req": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "description": "สำหรับ item ที่มีอยู่แล้ว",
                                "type": "integer"
                            },
                            "menu_item_id": {
                                "type": "integer"
                            },
                            "quantity": {
                                "type": "integer",
                                "minimum": 1
                            }
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "nameCh": {
                    "type": "string"
                },
                "nameEn": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}